Introduces hidden input handle property
Adds a `hidden` boolean property to `InputHandleBase` to allow certain inputs to be present in the node's schema but not rendered in the UI. This is useful for inputs that are expected to be set programmatically by other nodes rather than directly by the user.

Refactors the `ChatInput` node to include a new, hidden `message_in` input. This enables other nodes to provide the initial message to `ChatInput`, enhancing its flexibility.

Includes frontend support for the `hidden` property on `TextFieldHandleInput`.

Also adds a new `ImageInputInputHandle` type.